chore(dumper): regroup cli commands
